# openapi.yaml
openapi: 3.0.3
info:
  title: Vision Financial API
  version: "1.0"
  description: >
    Self-hosted financial transaction manager.
    All responses are wrapped in the ADR-026 envelope:
    `{ ok: true, data: T }` on success, `{ ok: false, error: {...} }` on failure.
servers:
  - url: http://localhost:3000
    description: Local development

# ─── Shared envelope & list wrappers ─────────────────────────────────────────
components:
  parameters:
...
    OptionalLimit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
      description: >
        Page size. Omit (together with offset) to receive the whole collection.
        Values above the endpoint's cap are clamped.
    OptionalOffset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: Rows to skip. Omit (together with limit) to receive the whole collection.
  schemas:
...
    Envelope:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        data:
          description: Present on success
        error:
          $ref: "#/components/schemas/ErrorPayload"
    ErrorPayload:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
...
    OptionalPaginationFields:
      type: object
      required: [total]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ─── Categories ─────────────────────────────────────────────────────────
...
    Tag:
      type: object
      required: [id, slug, is_active, created_at, updated_at]
      properties:
        id:
          type: integer
        slug:
          type: string
        color:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TagList:
      allOf:
        - $ref: "#/components/schemas/OptionalPaginationFields"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Tag"
    TagCreate:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
        color:
          type: string
          nullable: true
    TagUpdate:
      type: object
      properties:
        color:
          type: string
          nullable: true
        is_active:
          type: boolean
...
    WatchlistItem:
      type: object
      required: [id, name, asset_class, target_price, currency, created_at, updated_at]
      properties:
        id:
          type: integer
        name:
          type: string
        symbol:
          type: string
          nullable: true
        asset_class:
          type: string
          enum: [stock, etf, crypto, metals]
        target_price:
          type: number
        currency:
          type: string
        notes:
          type: string
          nullable: true
        price_provider_id:
          type: string
          nullable: true
        added_price:
          description: Live price snapshotted when the item was added (ADR-097 backtest).
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ─── AI ──────────────────────────────────────────────────────────────────
    # ai_conversations row. Column aliases are camelCase on the wire
    # (aiChatRepository's CONVERSATION_COLUMNS), not snake_case.
...
paths:
...
  /api/tags:
    get:
      summary: List tags
      operationId: getTags
      tags: [Tags]
      parameters:
        - in: query
          name: is_active
          schema:
            type: boolean
          description: Filter by active status (omit for all)
        - $ref: "#/components/parameters/OptionalLimit"
        - $ref: "#/components/parameters/OptionalOffset"
      responses:
        "200":
          description: Tag list (complete unless limit/offset are supplied)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TagList"
    post:
      summary: Create or reactivate tag
      operationId: createTag
      tags: [Tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreate"
      responses:
        "201":
          description: Created or reactivated tag
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Tag"
  /api/tags/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
    patch:
      summary: Update tag color or active status
      operationId: updateTag
      tags: [Tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdate"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Tag"
        "404":
          description: Tag not found
    delete:
      summary: Soft-delete tag (sets is_active=false)
      operationId: deleteTag
      tags: [Tags]
      responses:
        "200":
          description: The deactivated tag (soft delete — the row survives)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "404":
          description: Tag not found
...
  /api/watchlist:
    get:
      summary: List watchlist items
      operationId: getWatchlist
      tags: [Watchlist]
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/WatchlistItem"
    post:
      summary: Add symbol to watchlist
      operationId: addToWatchlist
      tags: [Watchlist]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol]
              properties:
                symbol:
                  type: string
                name:
                  type: string
                currency:
                  type: string
      responses:
        "201":
          description: Added item
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WatchlistItem"
  /api/watchlist/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
    get:
      summary: Get watchlist item
      operationId: getWatchlistItem
      tags: [Watchlist]
      responses:
        "200":
          description: Watchlist item with current price
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WatchlistItem"
    patch:
      summary: Update watchlist item
      operationId: updateWatchlistItem
      tags: [Watchlist]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    delete:
      summary: Remove from watchlist
      operationId: deleteWatchlistItem
      tags: [Watchlist]
      responses:
        "204":
          description: No Content

  # ─── Splits ───────────────────────────────────────────────────────────────────
...